requirements.txt
asgiref==3.5.2
Django==4.1
djangorestframework==3.13.1
filter==0.0.0.20200724
pytz==2022.1
sqlparse==0.4.2
